moving feed to common assets
