+++ Personal Diary/Min Marn/Min Marn Oo - PD.txt
@@ -55,4 +55,6 @@ Week 12
 
 Week 13
 - Restructuring the project files and transforming the website from using different pages to single page application
-- How to host Microsoft SQL database on AWS RDS
+- Setting up SQL database on Microsoft Azure
+- Learning how to POST using fetch API and sharing the knowledge to group members
+- Setting up documentations for frontend and backend integrations
